Separate PlayerController into Private and Public controllers
